Removed bioler plate code
