+++ drawroute/src/main/java/com/maps/route/utils/MapUtils.kt
@@ -1,0 +1,28 @@
+package com.maps.route.utils
+
+import android.graphics.Bitmap
+import android.graphics.Canvas
+import android.graphics.drawable.Drawable
+import com.google.android.gms.maps.model.*
+
+object MapUtils {
+
+    /**
+     * Helper method to get BitmapDescriptor from drawable
+     * @return void
+     * @param drawable Drawable resource
+     * @author Dawar Malik.
+     */
+    fun getMarkerIconFromDrawable(drawable: Drawable?): BitmapDescriptor {
+        val canvas = Canvas()
+        val bitmap = Bitmap.createBitmap(
+            drawable!!.intrinsicWidth,
+            drawable.intrinsicHeight,
+            Bitmap.Config.ARGB_8888
+        )
+        canvas.setBitmap(bitmap)
+        drawable.setBounds(0, 0, drawable.intrinsicWidth, drawable.intrinsicHeight)
+        drawable.draw(canvas)
+        return BitmapDescriptorFactory.fromBitmap(bitmap)
+    }
+}
